requirements.txt
celery==5.0.5
dj-inmemorystorage==2.1.0
Django==3.1.6
django-celery-results==2.0.1
django-sslserver==0.22
django-widget-tweaks==1.4.8
google-cloud-vision==2.3.1
itsdangerous==1.1.0
num2words==0.5.10
Pillow==8.1.2
pytest-cov==2.11.1
pytest-django==4.1.0
python-dotenv==0.15.0
gunicorn
psycopg2
